+++ SwiftySCAD/SwiftySCADFramework/3DShapes.swift
@@ -7,25 +7,25 @@ import Foundation
 //
 // cylinder(r = 3.5, h = 1, center = true, $fn = 32);
 //
-func cylinder(radius radius: Float, height: Float, resolution: Int = 32) -> String {
+func cylinder(radius radius: Double, height: Double, resolution: Int = 32) -> String {
     return "cylinder(r = \(radius), h = \(height), center = true, $fn = \(resolution));"
 }
 
-func cylinder(diameter diameter: Float, height: Float, resolution: Int = 32) -> String {
+func cylinder(diameter diameter: Double, height: Double, resolution: Int = 32) -> String {
     return cylinder(radius: diameter/2.0, height: height, resolution: resolution)
 }
 
 //
 // cylinder(h = height, r1 = BottomRadius, r2 = TopRadius, center = true/false);
 //
-func cylinder(topDiameter topDiameter: Float, bottomDiameter: Float, height: Float) -> String {
+func cylinder(topDiameter topDiameter: Double, bottomDiameter: Double, height: Double) -> String {
     return "cylinder(h = \(height), r1 = \(bottomDiameter/2.0), r2 = \(topDiameter/2.0), center = false);"
 }
 
 //
 // cube([width + height, length, height], center = true);
 //
-func cube(x x: Float, y: Float, z: Float) -> String {
+func cube(x x: Double, y: Double, z: Double) -> String {
     return "cube([\(x), \(y), \(z)], center = true);"
 }
 
@@ -34,7 +34,7 @@ func cube(x x: Float, y: Float, z: Float) -> String {
 //
 // module trapezoid(width_base, width_top,height,thickness)
 //
-func trapezoid(topWidth topWidth: Float, bottomWidth: Float, height: Float, thickness: Float) -> String {
+func trapezoid(topWidth topWidth: Double, bottomWidth: Double, height: Double, thickness: Double) -> String {
     let halfBottom = bottomWidth/2.0
     let halfTop = topWidth/2.0
     
@@ -53,7 +53,7 @@ func trapezoid(topWidth topWidth: Float, bottomWidth: Float, height: Float, thic
 //
 // width: flat-side to flat-side width measurement
 //
-func hexagon(width width: Float, thickness: Float) -> String {
+func hexagon(width width: Double, thickness: Double) -> String {
     let circleRadius = width.half/(sqrt(3.0)*0.5)
     
     let circle = "linear_extrude(height = \(thickness)) circle(r= \(circleRadius), $fn=6);"
